some bugs fixed and confirm modal shown upon delete

+++ src/components/frontDesign/EditCategory/EditCategory.jsx
@@ -6,6 +6,7 @@ import { BiEdit, BiTrash } from "react-icons/bi";
 import { PlusOutlined } from "@ant-design/icons";
 import { toast } from "react-toastify";
 import IsAdmin from "@/components/common/IsAdmin";
+import Swal from "sweetalert2";
 
 const EditCategory = ({ refetch, id, data }) => {
   const [modalIsOpen, setIsOpen] = useState(false);
@@ -70,7 +71,7 @@ const EditCategory = ({ refetch, id, data }) => {
   const handleSubmit = async () => {
     try {
       let imageUrl = categoryData.thumbnail;
-      
+
       if (fileList.length > 0 && fileList[0].originFileObj) {
         const imgData = new FormData();
         imgData.append("image", fileList[0].originFileObj);
@@ -142,26 +143,38 @@ const EditCategory = ({ refetch, id, data }) => {
   };
 
   const handleDeleteItem = async () => {
-    try {
-      const result = await fetch(`/api/category?id=${data._id}`, {
-        method: "DELETE",
-        headers: {
-          "Content-Type": "application/json",
-        },
-      });
-
-      const resultData = await result.json();
-
-      if (resultData.success) {
-        refetch();
-        toast.success(resultData.message);
-      } else {
-        toast.error(resultData.message);
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        try {
+          fetch(`/api/category?id=${data._id}`, {
+            method: "DELETE",
+            headers: {
+              "Content-Type": "application/json",
+            },
+          })
+            .then((res) => res.json())
+            .then((resultData) => {
+              if (resultData.success) {
+                refetch();
+                toast.success(resultData.message);
+              } else {
+                toast.error(resultData.message);
+              }
+            });
+        } catch (error) {
+          toast.error("Something went wrong!");
+          console.error("Error uploading image: ", error);
+        }
       }
-    } catch (error) {
-      toast.error("Something went wrong!");
-      console.error("Error uploading image: ", error);
-    }
+    });
   };
 
   return (
